+++ .config/hypr/hyprland.conf
@@ -159,7 +159,7 @@ misc {
 #-----------------------------------------------------------------------------------------------------------
 
 render {
-    explicit_sync = true
+    explicit_sync = false
 }
 
 #-----------------------------------------------------------------------------------------------------------
