+++ ignis/modules/bar/widgets/__init__.py
@@ -1,0 +1,8 @@
+from .pill import StatusPill
+from .tray import Tray
+from .kb_layout import KeyboardLayout
+from .battery import Battery
+from .apps import Apps
+from .workspaces import Workspaces
+
+__all__ = ["StatusPill", "Tray", "KeyboardLayout", "Battery", "Apps", "Workspaces"]
