+++ ignis/modules/bar/widgets/apps.py
@@ -33,19 +33,18 @@ class AppItem(Widget.Button):
         )
 
 
-def launcher_button():
-    return Widget.Button(
-        child=Widget.Icon(image="start-here-symbolic", pixel_size=32),
-        on_click=lambda x: app.toggle_window("ignis_LAUNCHER"),
-        css_classes=["pinned-app", "unset"],
-    )
-
-
-def pinned_apps():
-    return Widget.Box(
-        child=applications.bind(
-            "pinned",
-            transform=lambda value: [AppItem(app) for app in value]
-            + [launcher_button()],
+class Apps(Widget.Box):
+    def __init__(self):
+        super().__init__(
+            child=applications.bind(
+                "pinned",
+                transform=lambda value: [AppItem(app) for app in value]
+                + [
+                    Widget.Button(
+                        child=Widget.Icon(image="start-here-symbolic", pixel_size=32),
+                        on_click=lambda x: app.toggle_window("ignis_LAUNCHER"),
+                        css_classes=["pinned-app", "unset"],
+                    )
+                ],
+            )
         )
-    )
