+++ ignis/modules/bar/widgets/workspaces.py
@@ -3,6 +3,7 @@ from ignis.services.hyprland import HyprlandService
 
 hyprland = HyprlandService.get_default()
 
+
 class WorkspaceButton(Widget.Button):
     def __init__(self, workspace: dict) -> None:
         super().__init__(
@@ -13,8 +14,6 @@ class WorkspaceButton(Widget.Button):
         )
         if workspace["id"] == hyprland.active_workspace["id"]:
             self.add_css_class("active")
-
-
 
 
 def scroll_workspaces(direction: str) -> None:
@@ -29,14 +28,20 @@ def scroll_workspaces(direction: str) -> None:
         hyprland.switch_to_workspace(target)
 
 
-def workspaces():
-    return Widget.EventBox(
-        on_scroll_up=lambda x: scroll_workspaces("up"),
-        on_scroll_down=lambda x: scroll_workspaces("down"),
-        css_classes=["workspaces"],
-        visible=hyprland.is_available,
-        child=hyprland.bind(
-            "workspaces",
-            transform=lambda value: [WorkspaceButton(i) for i in value],
-        ),
-    )
+class Workspaces(Widget.Box):
+    def __init__(self):
+        if hyprland.is_available:
+            child = [
+                Widget.EventBox(
+                    on_scroll_up=lambda x: scroll_workspaces("up"),
+                    on_scroll_down=lambda x: scroll_workspaces("down"),
+                    css_classes=["workspaces"],
+                    child=hyprland.bind(
+                        "workspaces",
+                        transform=lambda value: [WorkspaceButton(i) for i in value],
+                    ),
+                )
+            ]
+        else:
+            child = []
+        super().__init__(child=child)
