+++ ignis/modules/control_center/quick_settings/__init__.py
@@ -1,0 +1,3 @@
+from .quick_settings import quick_settings
+
+__all__ = ["quick_settings"]
