+++ ignis/modules/shared_widgets/__init__.py
@@ -1,0 +1,3 @@
+from .toggle_box import ToggleBox
+
+__all__ = ["ToggleBox"]
